+++ packages/ui/app/earn/page.tsx
@@ -1,138 +1,11 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import { base, mode } from 'viem/chains';
 
-import CommingSoon from '../_components/earn/CommingSoon';
-import EarnRows, { type EarnRow } from '../_components/earn/EarnRows';
+import EarnRows from '../_components/earn/EarnRows';
 
-const earnOpps: EarnRow[] = [
-  {
-    apr: 0,
-    asset: ['ION', 'WETH'],
-    getApr: () => Promise.resolve(0),
-    getTvl: () => Promise.resolve(0),
-    rewards: {
-      [mode.id]: {
-        peaks: false
-      }
-    },
-    link: 'https://app.ionic.money/stake?chain=34443&token=weth',
-    network: 'mode',
-    poolChain: mode.id,
-    protocol: 'Velodrome',
-    tvl: 0,
-    tvlpool: '0xC6A394952c097004F83d2dfB61715d245A38735a'
-  },
-  {
-    apr: 0,
-    asset: ['ION', 'WETH'],
-    getApr: () => Promise.resolve(0),
-    getTvl: () => Promise.resolve(0),
-    rewards: {
-      [base.id]: {
-        peaks: false
-      }
-    },
-    link: 'https://app.ionic.money/stake?chain=8453',
-    network: 'base',
-    poolChain: base.id,
-    protocol: 'Aerodrome Finance',
-    tvl: 0,
-    tvlpool: '0x0FAc819628a7F612AbAc1CaD939768058cc0170c'
-  },
-  {
-    apr: 0,
-    asset: ['ION', 'MODE'],
-    getApr: () => Promise.resolve(0),
-    getTvl: () => Promise.resolve(0),
-    rewards: {
-      [mode.id]: {
-        peaks: false
-      }
-    },
-    link: 'https://app.ionic.money/stake?chain=34443&token=mode',
-    network: 'mode',
-    poolChain: mode.id,
-    protocol: 'Velodrome',
-    tvl: 0,
-    tvlpool: '0x690A74d2eC0175a69C0962B309E03021C0b5002E'
-  },
-  {
-    apr: 0,
-    asset: ['ionUSDC', 'ionUSDT'],
-    getApr: async () => {
-      try {
-        const response = await fetch(
-          'https://api.steer.finance/pool/fee-apr?address=0x17694615caba46ef765a3673fa488e04332b522a&chain=34443&interval=604800'
-        );
-        if (!response.ok) {
-          throw new Error(`HTTP error: Status ${response.status}`);
-        }
-        const val = await response.json();
-        return val?.apr ?? 0;
-      } catch (err) {
-        console.error(err);
-      }
-    },
-    getTvl: async () => {
-      try {
-        const response = await fetch(
-          'https://api.steer.finance/pool/lp/value?chain=34443&address=0x17694615caba46ef765a3673fa488e04332b522a'
-        );
-        if (!response.ok) {
-          throw new Error(`HTTP error: Status ${response.status}`);
-        }
-        const val = await response.json();
-        return val?.tvl ?? 0;
-      } catch (err) {
-        console.error(err);
-      }
-    },
-    rewards: {
-      [mode.id]: {
-        peaks: false
-      }
-    },
-    link: 'https://app.steer.finance/vault/0x17694615caba46ef765a3673fa488e04332b522a/34443',
-    network: 'mode',
-    protocol: 'Steer',
-    tvl: 0,
-    poolChain: mode.id
-  },
-  {
-    apr: 0,
-    asset: ['ionUSDC'],
-    getApr: () => Promise.resolve(0),
-    getTvl: () => Promise.resolve(0),
-    rewards: {
-      [mode.id]: {
-        peaks: true
-      }
-    },
-    link: 'https://davos.xyz/app/loans/mint/?network=mode&token=ionUSDC',
-    network: 'mode',
-    protocol: 'Davos',
-    tvl: 0,
-    poolChain: mode.id
-  },
-  {
-    apr: 0,
-    asset: ['ionUSDT'],
-    getApr: () => Promise.resolve(0),
-    getTvl: () => Promise.resolve(0),
-    rewards: {
-      [mode.id]: {
-        peaks: true
-      }
-    },
-    link: 'https://davos.xyz/app/loans/mint/?network=mode&token=ionUSDT',
-    network: 'mode',
-    protocol: 'Davos',
-    tvl: 0,
-    poolChain: mode.id
-  }
-];
+import type { EarnRow } from '@ui/utils/earnUtils';
+import { earnOpps } from '@ui/utils/earnUtils';
 
 export default function Earn() {
   const [rows, setRows] = useState<EarnRow[]>(earnOpps);
@@ -183,7 +56,8 @@ export default function Earn() {
               link,
               poolChain,
               tvlpool,
-              rewards
+              rewards,
+              live
             },
             idx
           ) => (
@@ -198,17 +72,18 @@ export default function Earn() {
               link={link}
               key={idx}
               rewards={rewards}
+              live={live}
             />
           )
         )}
-        <CommingSoon
+        {/* <CommingSoon
           linktoProtocol={'https://www.tren.finance'}
           additionalText={'Tren Finance'}
         />
         <CommingSoon
           linktoProtocol={'https://peapods.finance'}
           additionalText={'Peapods Finance'}
-        />
+        /> */}
       </div>
     </>
   );
